requirements.txt
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
